+++ src/physical-layer/default-config.service.js
@@ -46,7 +46,7 @@
         */
 
         return {
-            CONSTELLATION_DIAGRAM_DECIBEL_LIMIT: 40,
+            CONSTELLATION_DIAGRAM_DECIBEL_LIMIT: -40,
             MINIMUM_POWER_DECIBEL: -99,
             FAKE_NOISE_MAX_AMPLITUDE: 0.001,
             CHANNEL_1_FREQUENCY: 1070,